+++ src/App.tsx
@@ -4,9 +4,8 @@ import routes from './routes';
 import { useSelector } from 'react-redux';
 import { RootState } from './redux/store';
 
-import SnackBar from './components/@shared/SnackBar';
-import NavBar from './components/NavBar';
-
+import { SnackBar } from './components/@shared';
+import { NavBar } from './components';
 import {
   Cart,
   Login,
