+++ src/components/@shared/Empty/index.tsx
@@ -1,5 +1,5 @@
-import emptyPage from '../../../assets/emptyPage.jpeg';
 import { EmptyImageWrapper } from './styles';
+import emptyPage from '@/assets/emptyPage.jpeg';
 
 function Empty() {
   return (
