+++ src/components/CartProductList/index.tsx
@@ -1,11 +1,14 @@
 import { useDispatch } from 'react-redux';
-import { CartItem, deleteBySelectedItems, selectAllItems } from '../../redux/modules/cart';
-import { useCartListSelector } from '../../hooks/useCartSelector';
+import { CartItem, deleteBySelectedItems, selectAllItems } from '@/redux/modules/cart';
 
-import { Button, CheckBox } from '../@shared';
-import CartProduct from '../CartProduct';
-import { INFO_MESSAGES } from '../../constants';
+import { useCartListSelector } from '@/hooks/useCartSelector';
+
 import { CartListTitle, SelectAllContainer } from './styles';
+
+import { Button, CheckBox } from '@/components/@shared';
+import { CartProduct } from '@/components';
+
+import { INFO_MESSAGES } from '@/constants';
 
 function CartProductList() {
   const dispatch = useDispatch();
