+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { default as CartProduct } from './CartProduct';
+export { default as CartProductList } from './CartProductList';
+export { default as NavBar } from './NavBar';
+export { default as PageLayout } from './PageLayout';
+export { default as PaymentBox } from './PaymentBox';
+export { default as Product } from './Product';
