+++ src/pages/ProductList/index.tsx
@@ -1,10 +1,12 @@
 import { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { loadProductsAPI, ProductState, selectProductState } from '../../redux/modules/products';
 
-import { Loader } from '../../components/@shared';
-import Product from '../../components/Product';
+import { useSelector, useDispatch } from 'react-redux';
+import { loadProductsAPI, ProductState, selectProductState } from '@/redux/modules/products';
+
 import { ProductListContainer } from './styles';
+
+import { Loader } from '@/components/@shared';
+import Product from '@/components/Product';
 
 function ProductList() {
   const { productList, loading, error }: ProductState = useSelector(selectProductState);
